+++ src/Domain/Entities/User.cs
@@ -12,5 +12,5 @@ public class User: BaseAuditableEntity
     public string? Address { get; set; }
     public string? Phone { get; set; }
     public UserType? UserType { get; set; }
-    public decimal? Money { get; set; }
+    public float? Money { get; set; }
 }
